fix warnings; update chapter four

+++ harness/src/test/scala/com/robertjneal/rl/AgentTest.scala
@@ -51,11 +51,11 @@ class AgentTest {
     assertEquals(result1.e.possibleStateActions, result2.e.possibleStateActions)
     assertEquals(
       2,
-      result2.actionSteps(OneState).values.map(_.toInt).sum - actionRewards.size
+      result2.actionSteps(OneState).values.map(_.toInt).sum - actionRewards.size.toLong
     )
     assertEquals(
       result2.step,
-      result2.actionSteps(OneState).values.map(_.toInt).sum - actionRewards.size
+      result2.actionSteps(OneState).values.map(_.toInt).sum - actionRewards.size.toLong
     )
   }
 }
